+++ src/front-door/azext_front_door/commands.py
@@ -89,6 +89,9 @@ def load_command_table(self, _):
     #     g.show_command('show', 'get')
     #     g.generic_update_command('update', custom_func_name='update_fd_backend_pool')
 
+    with self.command_group('network front-door probe', frontdoor_sdk) as g:
+        g.custom_command('update', 'update_fd_health_probe_settings')
+
     with self.command_group('network front-door backend-pool backend', frontdoor_sdk) as g:
         g.custom_command('add', 'add_fd_backend')
         g.custom_command('list', 'list_fd_backends')