+++ src/hack/HISTORY.rst
@@ -1,7 +1,8 @@
 .. :changelog:
 
 Release History
+===============
 
 0.1.0
-+++++
-* Alpha version
+++++++
+* Initial release.