+++ src/healthcareapis/HISTORY.rst
@@ -3,6 +3,18 @@
 Release History
 ===============
 
+0.1.2
++++++
+
+* Upgraded API version to latest stable
+* Removed preview tag
+
+0.1.1
++++++
+
+* Updated examples
+* Fixed bugs
+
 0.1.0
 ++++++
 * Initial release.